requirements.txt
pytest
allure-pytest
reportportal-client
requests
azure-devops
pytest
